revert to a version prior (#499)
* revert to a version prior

* Cancel rsvp (#478)

* cancel rspved contest

* refactor

* add unit test

* use Repo.delete instead of Repo.delete_all

* Update lib/quadquizaminos/contests.ex
