--- lib/quadquizaminos_web/live/admin_contests_live.html.leex
@@ -1,27 +1,0 @@
-    <div class="container">
-        <h1>Contests</h1>
-         <input type="text" phx-keydown="save"  phx-key="Enter">
-             <table>
-            <tr>
-                <th></th><!-- contest name -->
-                <th></th> <!--start pause button-->
-                <th></th><!-- STOP button-->
-                <th></th><!--count up timer-->
-                <th>Date</th>
-                <th>Start</th>
-                <th>End</th>
-            </tr>
-
-            <%= for {status, contests} <- @contests |> group_contest_by_status do %>
-              <tr>
-                <td colspan="8"><b><%= status(status)  %> Contests</b></td>
-              </tr>
-                  <%= for  contest <- contests do %>
-                    <tr>
-                    <%= live_component @socket, QuadquizaminosWeb.ContestsLive.AdminContestComponent, id: contest.id, current_user: @current_user %>
-                    </tr>
-                    <% end %>
-            <% end %>
-        </table>
-        
-</div>
