+++ lib/quadquizaminos_web/live/contests_live.ex
@@ -3,16 +3,17 @@ defmodule QuadquizaminosWeb.ContestsLive do
 
   import Phoenix.LiveView.Helpers
   import Phoenix.HTML, only: [raw: 1]
-  import Phoenix.HTML.{Form, Tag}
-  alias QuadquizaminosWeb.AdminContestsLive
   alias Quadquizaminos.Contests
   alias QuadquizaminosWeb.ContestsLive.ContestComponent
   alias Quadquizaminos.Accounts
   alias Quadquizaminos.Accounts.User
   alias Quadquizaminos.Util
 
+  @conference_date Application.fetch_env!(:quadquizaminos, :conference_date)
+
   def mount(_params, session, socket) do
     :timer.send_interval(1000, self(), :update_component_timer)
+    countdown_interval = DateTime.diff(@conference_date, DateTime.utc_now())
     QuadquizaminosWeb.Endpoint.subscribe("contest_scores")
     current_user = session["uid"] |> current_user()
 
@@ -21,9 +22,59 @@ defmodule QuadquizaminosWeb.ContestsLive do
      |> assign(
        current_user: current_user,
        contests: Contests.list_contests(),
+       countdown_interval: countdown_interval,
        contest_records: [],
-       contest_id: nil
+       contest_id: nil,
+       editing_date?: false
      )}
+  end
+
+  def handle_event("add_contest_date", %{"contest" => name}, socket) do
+    contests =
+      Enum.map(socket.assigns.contests, fn contest ->
+        if contest.name == name do
+          %{contest | add_contest_date: true}
+        else
+          contest
+        end
+      end)
+
+    {:noreply, assign(socket, contests: contests, editing_date?: true)}
+  end
+
+  def handle_event("edit_contest_date", %{"contest" => name}, socket) do
+    contests =
+      Enum.map(socket.assigns.contests, fn contest ->
+        if contest.name == name do
+          %{contest | editing_date?: true}
+        else
+          contest
+        end
+      end)
+
+    {:noreply, assign(socket, contests: contests, editing_date?: true)}
+  end
+
+  def handle_event("save", %{"key" => "Enter", "value" => contest_name}, socket) do
+    {:noreply, socket |> _create_contest(contest_name)}
+  end
+
+  def handle_event("start", %{"contest" => name}, socket) do
+    {:noreply, start_or_resume_contest(socket, name)}
+  end
+
+  def handle_event("restart", %{"contest" => name}, socket) do
+    {:noreply, _restart_contest(socket, name)}
+  end
+
+  def handle_event("stop", %{"contest" => name}, socket) do
+    QuadquizaminosWeb.Endpoint.broadcast(
+      "contest_record",
+      "record_contest_scores",
+      name
+    )
+
+    {:noreply, _end_contest(socket, name)}
   end
 
   def handle_info(:update_component_timer, socket) do
@@ -39,8 +90,9 @@ defmodule QuadquizaminosWeb.ContestsLive do
     {:noreply, socket}
   end
 
-  def handle_info(:timer, socket) do
-    {:noreply, AdminContestLive._update_contests_timer(socket)}
+  def handle_info(:count_down, socket) do
+    countdown_interval = DateTime.diff(@conference_date, DateTime.utc_now())
+    {:noreply, socket |> assign(countdown_interval: countdown_interval)}
   end
 
   def handle_info(%{event: "current_scores", payload: payload}, socket) do
@@ -83,48 +135,114 @@ defmodule QuadquizaminosWeb.ContestsLive do
     end
   end
 
-  defp timer_or_final_result(assigns, contest) do
-    if contest.end_time do
-      ~L"""
-      <%= live_redirect "Final Results", class: "button",  to: Routes.contests_path(@socket, :show, contest)%>
+  defp _create_contest(socket, contest_name) do
+    contests = socket.assigns.contests
 
-      """
-    else
-      ~L"""
-
-      <% {hours, minutes, seconds} = contest.time_elapsed |> to_human_time() %>
-      <p><%= Util.count_display(hours) %>:<%= Util.count_display(minutes) %>:<%= Util.count_display(seconds) %></p>
-
-      """
+    case Contests.create_contest(%{name: contest_name}) do
+      {:ok, contest} -> assign(socket, contests: contests ++ [contest])
+      _ -> socket
     end
   end
 
-  defp live_result_button(assigns, contest) do
-    if contest.name in Contests.active_contests_names() do
-      ~L"""
-      <%= live_redirect "Live Results", class: "button",  to: Routes.contests_path(@socket, :show, contest)%>
+  defp _start_contest(socket, name) do
+    contests =
+      Enum.map(socket.assigns.contests, fn
+        contest ->
+          if contest.name == name do
+            {:ok, {:ok, started_contest}} = Contests.start_contest(name)
+            started_contest
+          else
+            contest
+          end
+      end)
 
-      """
+    assign(socket, contests: contests)
+  end
+
+  defp _restart_contest(socket, name) do
+    contests =
+      Enum.map(socket.assigns.contests, fn
+        contest ->
+          if contest.name == name do
+            {:ok, restarted_contest} = Contests.restart_contest(name)
+            restarted_contest
+          else
+            contest
+          end
+      end)
+
+    assign(socket, contests: contests)
+  end
+
+  defp _end_contest(socket, name) do
+    contests =
+      Enum.map(socket.assigns.contests, fn
+        contest ->
+          if contest.name == name do
+            {:ok, {:ok, ended_contest}} = Contests.end_contest(name)
+            ended_contest
+          else
+            contest
+          end
+      end)
+
+    assign(socket, contests: contests)
+  end
+
+  defp display_text_input(true = _admin?) do
+    """
+    <input type="text" phx-keydown="save"  phx-key="Enter">
+    """
+  end
+
+  defp display_text_input(_), do: ""
+
+  defp start_or_resume_contest(socket, name) do
+    if name in Contests.active_contests_names() && GenServer.whereis(name |> String.to_atom()) do
+      Contests.resume_contest(name)
+      socket
     else
-      ""
+      _start_contest(socket, name)
     end
   end
 
-  defp to_human_time(seconds) do
-    hours = div(seconds, 3600)
-    rem = rem(seconds, 3600)
-    minutes = div(rem, 60)
-    rem = rem(rem, 60)
-    seconds = div(rem, 1)
-    {hours, minutes, seconds}
+  defp admin?(current_user) do
+    ids = Application.get_env(:quadquizaminos, :github_ids)
+
+    current_user in (ids |> Enum.map(&(&1 |> to_string())))
   end
 
-  def truncate_date(nil) do
-    nil
-  end
+  def countdown_timer(_assigns, true = _admin?), do: ""
 
-  def truncate_date(date) do
-    DateTime.truncate(date, :second)
+  def countdown_timer(assigns, _) do
+    ~L"""
+    <div class="container">
+            <section class="phx-hero">
+                <h1>Contest countdown </h1>
+                <div class="row">
+                    <div class="column column-25">
+                        <% {days, hours, minutes, seconds} = Util.to_human_time(@countdown_interval) %>
+
+                        <h2><%= days |> Util.count_display() %></h2>
+                        <h2>DAYS</h2>
+                    </div>
+                    <div class="column column-25">
+                        <h2><%= Util.count_display(hours)%></h2>
+                        <h2>HOURS</h2>
+                    </div>
+                    <div class="column column-25">
+                        <h2><%=Util.count_display(minutes)%></h2>
+                        <h2>MINUTES</h2>
+                    </div>
+                    <div class="column column-25">
+                        <h2><%=Util.count_display(seconds)%></h2>
+                        <h2>SECONDS</h2>
+                    </div>
+                </div>
+                <h3>Contest coming soon</h3>
+            </section>
+        </div>
+    """
   end
 
   defp group_contest_by_status(contests) do
@@ -151,11 +269,5 @@ defmodule QuadquizaminosWeb.ContestsLive do
       nil -> %User{uid: "anonymous", name: "anonymous"}
       %User{} = user -> %{user | admin?: admin?(uid)}
     end
-  end
-
-  defp admin?(current_user) do
-    ids = Application.get_env(:quadquizaminos, :github_ids)
-
-    current_user in (ids |> Enum.map(&(&1 |> to_string())))
   end
 end
