+++ lib/quadquizaminos_web/router.ex
@@ -65,7 +65,6 @@ defmodule QuadquizaminosWeb.Router do
     pipe_through [:browser, :authorize_admin]
 
     live "/", AdminLive
-    live "/contest", AdminContestsLive
   end
 
   scope "/auth", QuadquizaminosWeb do
