+++ lib/quadquizaminos_web/views/layout_view.ex
@@ -67,7 +67,7 @@ defmodule QuadquizaminosWeb.LayoutView do
     <i class="fa fa-caret-down"></i>
     </a>
      <div class="dropdown-content">
-     <a class="button" href="<%= Routes.live_path(QuadquizaminosWeb.Endpoint, QuadquizaminosWeb.AdminContestsLive) %>">
+     <a class="button" href="<%= Routes.contests_path(QuadquizaminosWeb.Endpoint, :index) %>">
      Contest
      </a>
      <a class = "button" href="<%= Routes.live_path(QuadquizaminosWeb.Endpoint, QuadquizaminosWeb.AdminLive) %>">
